feat(nuxt3): add mock for @vue/composition-api
