+++ packages/nuxt3/src/core/nuxt.ts
@@ -78,6 +78,7 @@ export async function loadNuxt (opts: LoadNuxtOptions): Promise<Nuxt> {
   options.buildModules.push(pagesModule, metaModule, componentsModule, autoImportsModule)
   options.modulesDir.push(resolve(pkgDir, 'node_modules'))
   options.alias['vue-demi'] = resolve(options.appDir, 'compat/vue-demi')
+  options.alias['@vue/composition-api'] = resolve(options.appDir, 'compat/capi')
 
   const nuxt = createNuxt(options)
